fix: hook in seperate folder & refactor names
